+++ src/components/organisms/HotKeysHandler/HotKeysHandler.tsx
@@ -23,9 +23,9 @@ import {
   kubeConfigPathSelector,
   settingsSelector,
 } from '@redux/selectors';
-import {applyFileWithConfirm} from '@redux/services/applyFileWithConfirm';
 import {isKustomizationResource} from '@redux/services/kustomize';
 import {startPreview, stopPreview} from '@redux/services/preview';
+import {applyFileWithConfirm} from '@redux/support/applyFileWithConfirm';
 import {applyResource} from '@redux/thunks/applyResource';
 import {selectFromHistory} from '@redux/thunks/selectionHistory';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
