+++ src/redux/support/replaceSelectedMatchesWithConfirm.tsx
